Tested protocol handling. Added some functionality to protocol handling. Performed initial end to end test and discovered some bugs.

+++ src/usn/service.core/service.usn.core/service.usn.bridge/src/main/java/game/usn/bridge/util/USNBridgeUtil.java
@@ -26,7 +26,7 @@ public class USNBridgeUtil
      * @throws IllegalArgumentException
      *             - throw {@link IllegalArgumentException} if port number not valid.
      */
-    public static void validateNetworkPort(short port) throws IllegalArgumentException
+    public static void validateNetworkPort(int port) throws IllegalArgumentException
     {
         if (port < 0 || port > 65535)
         {
